+++ src/index.ts
@@ -1,5 +1,3 @@
-export * from './asyncComputed.js';
-export * from './asyncEffect.js';
 export * from './computed.js';
 export * from './effect.js';
 export * from './effectScope.js';
