+++ src/unstable/asyncEffect.ts
@@ -1,5 +1,6 @@
-import { Effect, nextTrackId } from './effect.js';
-import { asyncCheckDirty, Dependency, endTrack, link, startTrack, SubscriberFlags } from './system.js';
+import { Effect, nextTrackId } from '../effect.js';
+import { Dependency, endTrack, link, startTrack, SubscriberFlags } from '../system.js';
+import { asyncCheckDirty } from './asyncSystem.js';
 
 export function asyncEffect<T>(fn: () => AsyncGenerator<Dependency, T>): AsyncEffect<T> {
 	const e = new AsyncEffect(fn);
