+++ src/unstable/index.ts
@@ -1,3 +1,6 @@
+export * from './asyncComputed.js';
+export * from './asyncEffect.js';
+export * from './asyncSystem.js';
 export * from './computedArray.js';
 export * from './computedSet.js';
 export * from './equalityComputed.js';
